Isolate DBM related code
