+++ tox.ini
@@ -15,6 +15,7 @@ python =
     3.11: py311
 
 [testenv]
+extras = aiodbm
 deps=
     coverage
     factory_boy
